+++ tests/testthat/test-with_nix.R
@@ -15,14 +15,12 @@ testthat::test_that("Testing `with_nix()` if Nix is installed", {
 
   # Suppress the warning related to generating an expression
   # for an old version of R
-  suppressWarnings(
-    rix(
-      r_ver = "3.5.3",
-      overwrite = TRUE,
-      project_path = path_subshell,
-      message_type = "quiet",
-      shell_hook = NULL
-    )
+  rix(
+    r_ver = "3.5.3",
+    overwrite = TRUE,
+    project_path = path_subshell,
+    message_type = "quiet",
+    shell_hook = NULL
   )
 
   out_subshell <- with_nix(